changed soft_skeleton.py to loss_utils.py
