+++ mist/runtime/loss.py
@@ -1,7 +1,7 @@
 import torch
 import torch.nn as nn
 from torch.nn.functional import softmax, one_hot
-from .soft_skeleton import SoftSkeletonize
+from mist.runtime.loss_utils import SoftSkeletonize
 
 def get_one_hot(y_true, n_classes):
     y_true = y_true.to(torch.int64)
@@ -32,9 +32,9 @@ class DiceLoss(nn.Module):
         return loss
 
 class SoftCLDice(nn.Module):
-    def __init__(self, iter_=3, smooth=1., exclude_background=False):
+    def __init__(self, iterations=3, smooth=1., exclude_background=False):
         super(SoftCLDice, self).__init__()
-        self.iter = iter_
+        self.iterations = iterations
         self.smooth = smooth
         self.soft_skeletonize = SoftSkeletonize(num_iter=10)
         self.exclude_background = exclude_background
@@ -52,9 +52,9 @@ class SoftCLDice(nn.Module):
 
 
 class SoftDiceCLDice(nn.Module):
-    def __init__(self, iter_=3, alpha=0.5, smooth=1., exclude_background=False):
+    def __init__(self, iterations=3, alpha=0.5, smooth=1., exclude_background=False):
         super(SoftDiceCLDice, self).__init__()
-        self.iter = iter_
+        self.iterations = iterations
         self.smooth = smooth
         self.alpha = alpha
         self.dice_loss = DiceLoss()
@@ -73,26 +73,6 @@ class SoftDiceCLDice(nn.Module):
         cl_dice = 1. - 2.0 * (tprec * tsens) / (tprec + tsens)
         return (1.0 - self.alpha) * dice + self.alpha * cl_dice
 
-class soft_dice_cldice(nn.Module):
-    def __init__(self, iter_=3, alpha=0.5, smooth = 1., exclude_background=False):
-        super(soft_dice_cldice, self).__init__()
-        self.iter = iter_
-        self.smooth = smooth
-        self.alpha = alpha
-        self.soft_skeletonize = SoftSkeletonize(num_iter=10)
-        self.exclude_background = exclude_background
-
-    def forward(self, y_true, y_pred):
-        if self.exclude_background:
-            y_true = y_true[:, 1:, :, :]
-            y_pred = y_pred[:, 1:, :, :]
-        dice = soft_dice(y_true, y_pred)
-        skel_pred = self.soft_skeletonize(y_pred)
-        skel_true = self.soft_skeletonize(y_true)
-        tprec = (torch.sum(torch.multiply(skel_pred, y_true))+self.smooth)/(torch.sum(skel_pred)+self.smooth)    
-        tsens = (torch.sum(torch.multiply(skel_true, y_pred))+self.smooth)/(torch.sum(skel_true)+self.smooth)    
-        cl_dice = 1.- 2.0*(tprec*tsens)/(tprec+tsens)
-        return (1.0-self.alpha)*dice+self.alpha*cl_dice
 
 
 class DiceCELoss(nn.Module):
